rename logging.py so doesn't circ import (#1572)
